refactored Extension and Theme Modules
