+++ extensions/hello/src/HelloExtension.php
@@ -13,9 +13,9 @@ class HelloExtension extends Extension
     /**
      * {@inheritdoc}
      */
-    public function load(App $app, array $config)
+    public function __construct(App $app, array $config)
     {
-        parent::load($app, $config);
+        parent::__construct($app, $config);
 
         $app->subscribe(new HelloListener());
