+++ extensions/installer/extension.php
@@ -8,11 +8,8 @@ return [
 
     'main' => function ($app, $config) {
 
-        $extension = new InstallerExtension();
-        $extension->setConfig($config);
-        $extension->load($app, $config);
+        return new InstallerExtension($app, $config);
 
-        return $extension;
     },
 
     'autoload' => [
