+++ extensions/installer/src/InstallerExtension.php
@@ -11,21 +11,22 @@ class InstallerExtension extends Extension
     /**
      * {@inheritdoc}
      */
-    public function load(App $app, array $config)
+    public function __construct(App $app, array $config)
     {
-        if ($app['config']['app.installer']) {
-
-            parent::load($app, $config);
-
-            $app->error(function (NotFoundHttpException $e) use ($app) {
-                return $app['response']->redirect('@installer/installer');
-            });
-
-            $app->on('system.loaded', function () use ($app) {
-                if ($locale = $app['request']->getPreferredLanguage()) {
-                    $app['translator']->setLocale($locale);
-                }
-            });
+        if (!$app['config']['app.installer']) {
+            return;
         }
+
+        parent::__construct($app, $config);
+
+        $app->error(function (NotFoundHttpException $e) use ($app) {
+            return $app['response']->redirect('@installer/installer');
+        });
+
+        $app->on('system.loaded', function () use ($app) {
+            if ($locale = $app['request']->getPreferredLanguage()) {
+                $app['translator']->setLocale($locale);
+            }
+        });
     }
 }
