+++ extensions/page/src/PageExtension.php
@@ -16,9 +16,9 @@ class PageExtension extends Extension
     /**
      * {@inheritdoc}
      */
-    public function load(App $app, array $config)
+    public function __construct(App $app, array $config)
     {
-        parent::load($app, $config);
+        parent::__construct($app, $config);
 
         $app->subscribe(new NodeListener);
