+++ extensions/system/src/Theme.php
@@ -3,17 +3,11 @@
 namespace Pagekit\System;
 
 use Pagekit\Application as App;
-use Pagekit\Module\Module;
 use Pagekit\View\Section\SectionManager;
 use Pagekit\View\ViewInterface;
 
 class Theme extends Module
 {
-    /**
-     * @var array
-     */
-    protected $parameters = [];
-
     /**
      * @var string
      */
@@ -22,18 +16,9 @@ class Theme extends Module
     /**
      * {@inheritdoc}
      */
-    public function load(App $app, array $config)
+    public function __construct(App $app, array $config)
     {
-        if ($this->getConfig('parameters.settings')) {
-
-            if (is_array($defaults = $this->getConfig('parameters.settings.defaults'))) {
-                $this->parameters = array_replace($this->parameters, $defaults);
-            }
-
-            if (is_array($settings = App::option("{$config['name']}:settings"))) {
-                $this->parameters = array_replace($this->parameters, $settings);
-            }
-        }
+        parent::__construct($app, $config);
 
         $app->on('system.site', function() use ($app) {
             $this->registerRenderer($app['sections'], $app['view']);
@@ -61,37 +46,6 @@ class Theme extends Module
     public function getLayout()
     {
         return $this->getPath().$this->layout;
-    }
-
-    /**
-     * Returns the theme's parameters.
-     *
-     * @param  mixed $key
-     * @param  mixed $default
-     * @return array
-     */
-    public function getParams($key = null, $default = null)
-    {
-        if (null === $key) {
-            return $this->parameters;
-        }
-
-        $array = $this->parameters;
-
-        if (isset($array[$key])) {
-            return $array[$key];
-        }
-
-        foreach (explode('.', $key) as $segment) {
-
-            if (!is_array($array) || !array_key_exists($segment, $array)) {
-                return $default;
-            }
-
-            $array = $array[$segment];
-        }
-
-        return $array;
     }
 
     /**
